feat: adjust difficulty automatically by player count

+++ Assets/Script/BossMonster/Projectiles/Phase2/FireSpirit.cs
@@ -12,7 +12,7 @@ public class FireSpirit : NetworkBehaviour
     float currentTime = 0;
     [Networked] private CustomTickTimer life { get; set; }
     public List<PlayerRef> playersHit = new();
-    public int damage = 90;
+    public float damage = 90;
     public float speed = 5;
     public float lifeSeconds = 3;
     private int dir = 1;
